+++ src/sql/run.py
@@ -9,7 +9,7 @@ from io import StringIO
 import prettytable
 import sqlalchemy
 import sqlparse
-
+import sql.connection
 from .column_guesser import ColumnGuesserMixin
 
 try:
@@ -166,12 +166,23 @@ class ResultSet(list, ColumnGuesserMixin):
         for row in self:
             yield dict(zip(self.keys, row))
 
-    @telemetry.log_call("data-frame")
-    def DataFrame(self):
+    @telemetry.log_call("data-frame", payload=True)
+    def DataFrame(self, payload):
         "Returns a Pandas DataFrame instance built from the result set."
         import pandas as pd
 
         frame = pd.DataFrame(self, columns=(self and self.keys) or [])
+        payload[
+            "connection_info"
+        ] = sql.connection.Connection.current._get_curr_connection_info()
+        return frame
+
+    @telemetry.log_call("polars-data-frame")
+    def PolarsDataFrame(self):
+        "Returns a Polars DataFrame instance built from the result set."
+        import polars as pl
+
+        frame = pl.DataFrame((tuple(row) for row in self), schema=self.keys)
         return frame
 
     @telemetry.log_call("pie")
@@ -397,6 +408,8 @@ def run(conn, sql, config, user_namespace):
         resultset = ResultSet(result, statement, config)
         if config.autopandas:
             return resultset.DataFrame()
+        elif config.autopolars:
+            return resultset.PolarsDataFrame()
         else:
             return resultset
         # returning only last result, intentionally