+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # CHANGELOG
 
-## 0.5.7dev
+## 0.6.1dev
+
+## 0.6.0 (2023-02-27)
+
+* [API Change] Drops support for old versions of IPython (removed imports from `IPython.utils.traitlets`)
+* [Feature] Adds `%%config SqlMagic.autopolars = True` ([#138](https://github.com/ploomber/jupysql/issues/138))
 
 * [Feature] Support new variable substitution as {{a}} format ([#137](https://github.com/ploomber/jupysql/pull/137))